+++ core/src/upgrade/either.rs
@@ -19,18 +19,12 @@
 // DEALINGS IN THE SOFTWARE.
 
 use crate::{
-    either::{EitherError, EitherFuture2, EitherName, EitherOutput},
+    either::{EitherFuture2, EitherName, EitherOutput},
     upgrade::{InboundUpgrade, OutboundUpgrade, UpgradeInfo},
 };
+use either::Either;
 
-/// A type to represent two possible upgrade types (inbound or outbound).
-#[derive(Debug, Clone)]
-pub enum EitherUpgrade<A, B> {
-    A(A),
-    B(B),
-}
-
-impl<A, B> UpgradeInfo for EitherUpgrade<A, B>
+impl<A, B> UpgradeInfo for Either<A, B>
 where
     A: UpgradeInfo,
     B: UpgradeInfo,
@@ -43,27 +37,27 @@ where
 
     fn protocol_info(&self) -> Self::InfoIter {
         match self {
-            EitherUpgrade::A(a) => EitherIter::A(a.protocol_info().into_iter()),
-            EitherUpgrade::B(b) => EitherIter::B(b.protocol_info().into_iter()),
+            Either::Left(a) => EitherIter::A(a.protocol_info().into_iter()),
+            Either::Right(b) => EitherIter::B(b.protocol_info().into_iter()),
         }
     }
 }
 
-impl<C, A, B, TA, TB, EA, EB> InboundUpgrade<C> for EitherUpgrade<A, B>
+impl<C, A, B, TA, TB, EA, EB> InboundUpgrade<C> for Either<A, B>
 where
     A: InboundUpgrade<C, Output = TA, Error = EA>,
     B: InboundUpgrade<C, Output = TB, Error = EB>,
 {
     type Output = EitherOutput<TA, TB>;
-    type Error = EitherError<EA, EB>;
+    type Error = Either<EA, EB>;
     type Future = EitherFuture2<A::Future, B::Future>;
 
     fn upgrade_inbound(self, sock: C, info: Self::Info) -> Self::Future {
         match (self, info) {
-            (EitherUpgrade::A(a), EitherName::A(info)) => {
+            (Either::Left(a), EitherName::A(info)) => {
                 EitherFuture2::A(a.upgrade_inbound(sock, info))
             }
-            (EitherUpgrade::B(b), EitherName::B(info)) => {
+            (Either::Right(b), EitherName::B(info)) => {
                 EitherFuture2::B(b.upgrade_inbound(sock, info))
             }
             _ => panic!("Invalid invocation of EitherUpgrade::upgrade_inbound"),
@@ -71,21 +65,21 @@ where
     }
 }
 
-impl<C, A, B, TA, TB, EA, EB> OutboundUpgrade<C> for EitherUpgrade<A, B>
+impl<C, A, B, TA, TB, EA, EB> OutboundUpgrade<C> for Either<A, B>
 where
     A: OutboundUpgrade<C, Output = TA, Error = EA>,
     B: OutboundUpgrade<C, Output = TB, Error = EB>,
 {
     type Output = EitherOutput<TA, TB>;
-    type Error = EitherError<EA, EB>;
+    type Error = Either<EA, EB>;
     type Future = EitherFuture2<A::Future, B::Future>;
 
     fn upgrade_outbound(self, sock: C, info: Self::Info) -> Self::Future {
         match (self, info) {
-            (EitherUpgrade::A(a), EitherName::A(info)) => {
+            (Either::Left(a), EitherName::A(info)) => {
                 EitherFuture2::A(a.upgrade_outbound(sock, info))
             }
-            (EitherUpgrade::B(b), EitherName::B(info)) => {
+            (Either::Right(b), EitherName::B(info)) => {
                 EitherFuture2::B(b.upgrade_outbound(sock, info))
             }
             _ => panic!("Invalid invocation of EitherUpgrade::upgrade_outbound"),