+++ core/src/upgrade/select.rs
@@ -19,9 +19,10 @@
 // DEALINGS IN THE SOFTWARE.
 
 use crate::{
-    either::{EitherError, EitherFuture2, EitherName, EitherOutput},
+    either::{EitherFuture2, EitherName, EitherOutput},
     upgrade::{InboundUpgrade, OutboundUpgrade, UpgradeInfo},
 };
+use either::Either;
 
 /// Upgrade that combines two upgrades into one. Supports all the protocols supported by either
 /// sub-upgrade.
@@ -64,7 +65,7 @@ where
     B: InboundUpgrade<C, Output = TB, Error = EB>,
 {
     type Output = EitherOutput<TA, TB>;
-    type Error = EitherError<EA, EB>;
+    type Error = Either<EA, EB>;
     type Future = EitherFuture2<A::Future, B::Future>;
 
     fn upgrade_inbound(self, sock: C, info: Self::Info) -> Self::Future {
@@ -81,7 +82,7 @@ where
     B: OutboundUpgrade<C, Output = TB, Error = EB>,
 {
     type Output = EitherOutput<TA, TB>;
-    type Error = EitherError<EA, EB>;
+    type Error = Either<EA, EB>;
     type Future = EitherFuture2<A::Future, B::Future>;
 
     fn upgrade_outbound(self, sock: C, info: Self::Info) -> Self::Future {