+++ protocols/dcutr/src/handler/relayed.rs
@@ -21,11 +21,12 @@
 //! [`ConnectionHandler`] handling relayed connection potentially upgraded to a direct connection.
 
 use crate::protocol;
+use either::Either;
 use futures::future::{BoxFuture, FutureExt};
 use instant::Instant;
-use libp2p_core::either::{EitherError, EitherOutput};
+use libp2p_core::either::EitherOutput;
 use libp2p_core::multiaddr::Multiaddr;
-use libp2p_core::upgrade::{self, DeniedUpgrade, NegotiationError, UpgradeError};
+use libp2p_core::upgrade::{DeniedUpgrade, NegotiationError, UpgradeError};
 use libp2p_core::ConnectedPoint;
 use libp2p_swarm::handler::{
     ConnectionEvent, DialUpgradeError, FullyNegotiatedInbound, FullyNegotiatedOutbound,
@@ -129,7 +130,7 @@ pub struct Handler {
     /// A pending fatal error that results in the connection being closed.
     pending_error: Option<
         ConnectionHandlerUpgrErr<
-            EitherError<protocol::inbound::UpgradeError, protocol::outbound::UpgradeError>,
+            Either<protocol::inbound::UpgradeError, protocol::outbound::UpgradeError>,
         >,
     >,
     /// Queue of events to return when polled.
@@ -246,8 +247,8 @@ impl Handler {
                 // the remote peer and results in closing the connection.
                 self.pending_error = Some(error.map_upgrade_err(|e| {
                     e.map_err(|e| match e {
-                        EitherError::A(e) => EitherError::A(e),
-                        EitherError::B(v) => void::unreachable(v),
+                        Either::Left(e) => Either::Left(e),
+                        Either::Right(v) => void::unreachable(v),
                     })
                 }));
             }
@@ -286,7 +287,7 @@ impl Handler {
             _ => {
                 // Anything else is considered a fatal error or misbehaviour of
                 // the remote peer and results in closing the connection.
-                self.pending_error = Some(error.map_upgrade_err(|e| e.map_err(EitherError::B)));
+                self.pending_error = Some(error.map_upgrade_err(|e| e.map_err(Either::Right)));
             }
         }
     }
@@ -296,9 +297,9 @@ impl ConnectionHandler for Handler {
     type InEvent = Command;
     type OutEvent = Event;
     type Error = ConnectionHandlerUpgrErr<
-        EitherError<protocol::inbound::UpgradeError, protocol::outbound::UpgradeError>,
+        Either<protocol::inbound::UpgradeError, protocol::outbound::UpgradeError>,
     >;
-    type InboundProtocol = upgrade::EitherUpgrade<protocol::inbound::Upgrade, DeniedUpgrade>;
+    type InboundProtocol = Either<protocol::inbound::Upgrade, DeniedUpgrade>;
     type OutboundProtocol = protocol::outbound::Upgrade;
     type OutboundOpenInfo = u8; // Number of upgrade attempts.
     type InboundOpenInfo = ();
@@ -306,7 +307,7 @@ impl ConnectionHandler for Handler {
     fn listen_protocol(&self) -> SubstreamProtocol<Self::InboundProtocol, Self::InboundOpenInfo> {
         match self.endpoint {
             ConnectedPoint::Dialer { .. } => {
-                SubstreamProtocol::new(upgrade::EitherUpgrade::A(protocol::inbound::Upgrade {}), ())
+                SubstreamProtocol::new(Either::Left(protocol::inbound::Upgrade {}), ())
             }
             ConnectedPoint::Listener { .. } => {
                 // By the protocol specification the listening side of a relayed connection
@@ -314,7 +315,7 @@ impl ConnectionHandler for Handler {
                 // the relayed connection opens a substream to the dialing side. (Connection roles
                 // and substream roles are reversed.) The listening side on a relayed connection
                 // never expects incoming substreams, hence the denied upgrade below.
-                SubstreamProtocol::new(upgrade::EitherUpgrade::B(DeniedUpgrade), ())
+                SubstreamProtocol::new(Either::Right(DeniedUpgrade), ())
             }
         }
     }
@@ -387,7 +388,7 @@ impl ConnectionHandler for Handler {
                 }
                 Err(e) => {
                     return Poll::Ready(ConnectionHandlerEvent::Close(
-                        ConnectionHandlerUpgrErr::Upgrade(UpgradeError::Apply(EitherError::A(e))),
+                        ConnectionHandlerUpgrErr::Upgrade(UpgradeError::Apply(Either::Left(e))),
                     ))
                 }
             }