+++ swarm/tests/swarm_derive.rs
@@ -368,10 +368,10 @@ fn generated_out_event_derive_debug() {
 
 #[test]
 fn custom_out_event_no_type_parameters() {
-    use libp2p_core::connection::ConnectionId;
     use libp2p_core::PeerId;
     use libp2p_swarm::{
-        ConnectionHandler, IntoConnectionHandler, NetworkBehaviourAction, PollParameters,
+        ConnectionHandler, ConnectionId, IntoConnectionHandler, NetworkBehaviourAction,
+        PollParameters,
     };
     use std::task::Context;
     use std::task::Poll;