+++ swarm/CHANGELOG.md
@@ -15,6 +15,8 @@
 - Remove type parameter from `PendingOutboundConnectionError` and `PendingInboundConnectionError`.
   These two types are always used with `std::io::Error`. See [PR 3272].
 
+- Remove `ConnectionId::new`. Manually creating `ConnectionId`s is now unsupported. See [PR 3327].
+
 - Remove `handler` field from `NetworkBehaviourAction::Dial`.
   Instead of constructing the handler early, you now get to pass a `ConnectionId`.
   `ConnectionId` are `Copy` and will be used throughout the entire lifetime of the connection to report events.
@@ -26,6 +28,7 @@
 [PR 3153]: https://github.com/libp2p/rust-libp2p/pull/3153
 [PR 3264]: https://github.com/libp2p/rust-libp2p/pull/3264
 [PR 3272]: https://github.com/libp2p/rust-libp2p/pull/3272
+[PR 3327]: https://github.com/libp2p/rust-libp2p/pull/3327
 [PR 3328]: https://github.com/libp2p/rust-libp2p/pull/3328
 
 # 0.41.1